Move the data transfer objects to inmanta.dto
inmanta/data/model.py held the pydantic DTOs, but it lived inside the
inmanta.data package whose __init__.py is the database access layer. Python
executes a package __init__ before any of its submodules, so anything that
needed a DTO also got asyncpg, sqlalchemy and the whole DAO. That is why the
compiler paid for the database to reach a handful of pydantic models.

Move them to inmanta.dto and keep inmanta.data.model as an explicit re-export
of it, so the historical location and the @stable_api contract still hold.
Importing the DTOs goes from 1529 to 932 modules. All 44 core modules that
imported inmanta.data.model now import inmanta.dto; only the shim refers to the
old path.

Three dependency inversions had to go with it, since the DTOs are the lowest
layer and were reaching up into the two highest:

- DiscoveredResourceInput.to_dao built a DAO object. It becomes
  data.DiscoveredResource.from_dto on the DAO side.
- The DTOs called protocol.common.json_encode to serialize themselves.
  custom_json_encoder and json_encode move to inmanta.util, next to the
  api_boundary_json_encoder they already delegate to, and protocol.common
  re-exports them. inmanta.util.make_attribute_hash no longer needs its
  function level import of the protocol layer.
- PRIMITIVE_SQL_TYPES was defined in the DAO but is a plain type alias with no
  database semantics. It moves to inmanta.types, re-exported from inmanta.data.

Two more pieces of the same problem, needed for any of the above to have an
effect on the compiler:

- protocol.methods and protocol.auth.policy_engine each reach the DAO in one
  place. Both are server side code paths in a package every client imports, so
  the import is deferred to call time. Properly relocating convert_environment
  needs ArgOption to resolve its getter lazily, which is out of scope here.
- inmanta/agent/__init__.py re-exported Agent from agent_new, so importing
  inmanta.agent.handler, which every module and the compiler do, pulled the
  scheduler and the server. The package __init__ is now empty; import Agent
  from inmanta.agent.agent_new.
- inmanta.env and inmanta.app imported the server bootloader and the database
  and metrics services at module scope for uses inside single functions.

The compiler still reaches the DAO through one remaining edge:
inmanta.agent.handler annotates and builds data.LogLine. LogLine is a
DataDocument, which is explicitly not stored in the database, so it belongs in
the foundation, but inmanta.dto already has an unrelated LogLine and resolving
that name clash is a design decision rather than a move.
