+++ src/inmanta/agent/executor.py
@@ -43,7 +43,7 @@ from inmanta.agent import resourcepool
 from inmanta.agent.handler import HandlerContext
 from inmanta.const import Change
 from inmanta.data import LogLine
-from inmanta.data.model import AttributeStateChange, ModuleSource, PipConfig
+from inmanta.dto import AttributeStateChange, ModuleSource, PipConfig
 from inmanta.env import PythonEnvironment
 from inmanta.resources import Id
 from inmanta.types import JsonType, ResourceIdStr, ResourceVersionIdStr
