+++ src/inmanta/module.py
@@ -53,11 +53,11 @@ import pydantic
 import yaml
 from pydantic import BaseModel, Field, NameEmail, StringConstraints, ValidationError, field_validator
 
-import inmanta.data.model
+import inmanta.util
 import packaging.requirements
 import packaging.utils
 import packaging.version
-from inmanta import RUNNING_TESTS, const, env, loader, plugins
+from inmanta import RUNNING_TESTS, const, dto, env, loader, plugins
 from inmanta.ast import CompilerException, LocatableString, Location, Namespace, Range, WrappingRuntimeException
 from inmanta.ast.blocks import BasicBlock
 from inmanta.ast.statements import BiStatement, DefinitionStatement, DynamicStatement, Statement
@@ -1418,7 +1418,7 @@ class RelationPrecedenceRule:
 
 
 @stable_api
-class ProjectPipConfig(inmanta.data.model.PipConfig):
+class ProjectPipConfig(dto.PipConfig):
     """
     :param index_url: one pip index url for this project.
     :param extra_index_url:  additional pip index urls for this project. This is generally only
@@ -1535,7 +1535,7 @@ class ProjectMetadata(Metadata, MetadataFieldRequires):
     ] = []
     agent_install_dependency_modules: bool = True
     pip: ProjectPipConfig = ProjectPipConfig()
-    environment_settings: dict[str, inmanta.data.model.EnvSettingType] | None = None
+    environment_settings: dict[str, dto.EnvSettingType] | None = None
 
     @field_validator("modulepath", mode="before")
     @classmethod
