+++ src/inmanta/server/services/databaseservice.py
@@ -27,7 +27,7 @@ from typing import Callable, Mapping, Optional
 import asyncpg
 
 from inmanta.data import start_engine, stop_engine
-from inmanta.data.model import DataBaseReport, ReportedStatus
+from inmanta.dto import DataBaseReport, ReportedStatus
 from inmanta.server import SLICE_DATABASE
 from inmanta.server import config as opt
 from inmanta.server import protocol
