+++ src/inmanta/server/services/dryrunservice.py
@@ -22,7 +22,7 @@ import uuid
 from typing import Optional, cast
 
 from inmanta import const, data
-from inmanta.data.model import DryRun, DryRunReport, ResourceDiff, ResourceDiffStatus
+from inmanta.dto import DryRun, DryRunReport, ResourceDiff, ResourceDiffStatus
 from inmanta.protocol import handle, methods, methods_v2
 from inmanta.protocol.exceptions import Conflict, NotFound
 from inmanta.resources import Id
