+++ src/inmanta/server/services/resourceservice.py
@@ -27,8 +27,10 @@ from asyncpg.connection import Connection
 from pydantic import ValidationError
 from tornado.httputil import url_concat
 
-from inmanta import const, data, util
-from inmanta.data import APILIMIT, InvalidSort, model
+from inmanta import const, data
+from inmanta import dto as model
+from inmanta import util
+from inmanta.data import APILIMIT, InvalidSort
 from inmanta.data.dataview import (
     DiscoveredResourceView,
     ResourceHistoryView,
@@ -36,7 +38,7 @@ from inmanta.data.dataview import (
     ResourcesInVersionView,
     ResourceView,
 )
-from inmanta.data.model import (
+from inmanta.dto import (
     LatestReleasedResource,
     ReleasedResourceDetails,
     Resource,
@@ -449,14 +451,14 @@ class ResourceService(protocol.ServerSlice):
             LOGGER.exception(error_msg)
             raise BadRequest(error_msg, {"validation_errors": e.errors()})
 
-        dao = discovered_resource.to_dao(env.id)
+        dao = data.DiscoveredResource.from_dto(discovered_resource, env.id)
         await dao.insert_with_overwrite()
 
     @handle(methods_v2.discovered_resource_create_batch, env="tid")
     async def discovered_resources_create_batch(
         self, env: data.Environment, discovered_resources: list[model.DiscoveredResourceInput]
     ) -> None:
-        dao_list = [res.to_dao(env.id) for res in discovered_resources]
+        dao_list = [data.DiscoveredResource.from_dto(res, env.id) for res in discovered_resources]
         await data.DiscoveredResource.insert_many_with_overwrite(dao_list)
 
     @handle(methods_v2.discovered_resources_get, env="tid")
